+++ src/main/java/com/sefaz/sistemaficha/repositories/FichaFuncionalRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.sefaz.sistemaficha.entities.FichaFuncional;
 
 @Repository
-public interface FichaFuncionalRepository extends JpaRepository<FichaFuncional, Integer>{
+public interface FichaFuncionalRepository extends JpaRepository<FichaFuncional, Long>{
 
 }
